refactor:  rename component folder generic-list to generic_list
